(feat) Add billable exemptions schema
