+++ packages/esm-billing-app/src/billable-exemption/billable-exemptions.component.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import BillingHeader from '../../billing-header/billing-header.component';
-import { BillableExemptionsViewer } from './billable-exemptions-viewer.component';
+import BillingHeader from '../billing-header/billing-header.component';
+import { BillableExemptionsViewer } from '../billable-services/billable-exemptions/billable-exemptions-viewer.component';
 
 export const BillableExemptions = () => {
   const { t } = useTranslation();
