+++ packages/esm-billing-app/src/types/index.ts
@@ -548,3 +548,19 @@ export type FacilityClaim = {
   externalId: string;
 };
 export type BillingPromptType = 'patient-chart' | 'billing-orders';
+
+export interface Schema {
+  services: {
+    all: Service[];
+    'program:HIV'?: Service[];
+    'program:TB'?: Service[];
+    'age<5'?: Service[];
+    'visitAttribute:prisoner'?: Service[];
+  };
+  commodities: Record<string, unknown>;
+}
+
+export interface Service {
+  concept: string;
+  description: string;
+}
