+++ packages/esm-billing-app/translations/en.json
@@ -5,6 +5,7 @@
   "addBill": "Add bill item(s)",
   "addCommodityChargeItem": "Add charge item",
   "addPaymentOptions": "Add payment option",
+  "addSchema": "Add Schema",
   "addServiceChargeItem": "Add charge service",
   "all": "All",
   "amount": "Amount",
@@ -73,7 +74,6 @@
   "clockingOut": "Clocking Out",
   "clockOut": "Clock Out",
   "close": "Close",
-  "copySchema": "Copy schema",
   "create": "Create",
   "createClaimError": "Create Claim error",
   "created": "Created",
@@ -94,6 +94,7 @@
   "editBillForm": "Edit Bill Form",
   "editChargeItem": "Edit charge item",
   "editItem": "Edit item",
+  "editSchema": "Edit Schema",
   "editServiceChargeItem": "Edit Service Charge Item",
   "eligibleBenefits": "Eligible benefits",
   "emptyClaimsHeader": "No Claims",
@@ -119,6 +120,7 @@
   "errorOnLine": "Error on line",
   "errorRetrievingHIESubscription": "Error retrieving HIE subscription",
   "exemptionCategory": "Exemption category",
+  "exemptionSchema": "Exemption Schema",
   "facility": "Facility",
   "failedBillPayment": "Bill payment failed",
   "failure": "Error loading intervensions",
@@ -130,7 +132,6 @@
   "guaranteeId": "Guarantee Id",
   "home": "Home",
   "identifier": "Identifier",
-  "importSchema": "Import schema",
   "inactive": "Inactive",
   "incompletePayment": "Incomplete payment",
   "initiatePay": "Initiate Payment",
@@ -140,6 +141,7 @@
   "interventionCode": "Intervention Code",
   "interventions": "Interventions",
   "interventionsOption": "Choose interventions",
+  "invalidJsonError": "Invalid JSON received for the schema.",
   "invoice": "Invoice",
   "invoiceError": "Invoice error",
   "isPatientExemptedLegend": "Is patient exempted from payment?",
@@ -156,8 +158,6 @@
   "makeclaims": "Make Claims",
   "makePreauthRequest": "Make Pre-auth Request",
   "makePreAuthRequests": "Make Pre Auth Request",
-  "maximizeEditor": "Maximize editor",
-  "minimizeEditor": "Minimize editor",
   "mpesaPayment": "MPESA Payment",
   "name": "Name",
   "navigateBack": "Navigate back",
@@ -174,6 +174,8 @@
   "noMatchingBillsToDisplay": "No matching bills to display",
   "noMatchingItemsToDisplay": "No matching items to display",
   "noResultsSuggestion": "We couldn't find any matching benefit.",
+  "noSchema": "No schema available",
+  "noSchemaExemption": "No schema available add exemption schema",
   "notSearchedState": "Please search for a patient in the input above",
   "overflowMenu": "Overflow menu",
   "package": "Package",
@@ -233,17 +235,16 @@
   "quantity": "Quantity",
   "referenceCodes": "Reference Codes",
   "referenceNumber": "Reference number",
-  "referenceSchema": "Reference schema",
   "refundItem": "Refund item",
   "regExp": "Regular expression",
   "rejected": "Rejected",
   "remove": "Remove",
-  "renderChanges": "Render changes",
   "requestPreAuth": "Pre Authorization",
   "requestPreAuthError": "Pre Authorization Error",
   "resetFilters": " Reset filters",
   "retryRequest": "Retry request",
   "saveAndClose": "Save & Close",
+  "saveError": "Failed to save schema.",
   "schemaEditor": "Exemptions schema editor",
   "schemaLoadError": "Error loading schema",
   "search": "Search",
@@ -298,15 +299,5 @@
   "visitTime": "Visit time",
   "visittype": "Visit Type",
   "waiveBill": "Waive Bill",
-  "yes": "Yes",
-  "noSchema": "No schema available",
-  "editSchema": "Edit Schema",
-  "exemptionsSchemaEditor": "Exemptions Schema Editor",
-  "exemptionSchema": "Exemption Schema",
-  "invalidJsonError": "Invalid JSON received for the schema.",
-  "saveSchena": "Save Exemption Schema",
-  "saveSchemaToServer": "Save schema to server",
-  "errorCreatingSchema": "Error creating schema",
-  "schemaCreated": "New schema created",
-  "noSchemaExemption" : "No schema available add exemption schema"
+  "yes": "Yes"
 }
